requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
Django==2.1.5
django-rest-auth==0.9.3
djangorestframework==3.9.0
flickrapi==2.4.0
idna==2.8
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
oauthlib==2.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.1.0
requests-toolbelt==0.8.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
virtualenv==16.2.0
wcwidth==0.1.7
